+++ src/PowerPlantChallenge.WebApi/Models/ProductionPlan.cs
@@ -4,14 +4,14 @@ namespace PowerPlantChallenge.WebApi.Models
 {
     public class ProductionPlan
     {
-        public ProductionPlan(double load, Fuels fuels, IEnumerable<PowerPlant> powerPlants)
+        public ProductionPlan(decimal load, Fuels fuels, IEnumerable<PowerPlant> powerPlants)
         {
             Load = load;
             Fuels = fuels;
             PowerPlants = powerPlants;
         }
 
-        public double Load { get; }
+        public decimal Load { get; }
         public Fuels Fuels { get; }
         public IEnumerable<PowerPlant> PowerPlants { get; }
     }
